feat(RedisChatService): addMessage 수정 및 메세지 포멧 추가
